requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
# autopep8 formatting
autopep8==2.3.1